feat(steps): addded step to create new folder
Added input field for adding a containing folder name

BREAKING CHANGE: New step!
