+++ component.tsx
@@ -1,0 +1,12 @@
+import useStyles from "./styles";
+import { T } from "./type";
+
+
+const  = ({ }: T) => {
+	const { classes } = useStyles();
+	
+	return (<></>);
+
+}
+
+export default ;
