+++ styles.ts
@@ -1,0 +1,9 @@
+import { createStyles } from '@mantine/core';
+
+const useStyles = createStyles((theme) => ({
+	'.': {
+		// your styles here
+	},
+}));
+
+export default useStyles;
